requirements.txt
# python version 3.12.7

aiocache==0.12.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
disnake==2.9.2
fastapi==0.112.0
fastapi-discord==0.2.6.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.0
multidict==6.1.0
ormar[postgresql,sqlite,mysql]==0.20.1
propcache==0.2.0
pydantic==2.5.3
python-cas==1.6.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.14.0
